% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-doi.R
\name{check_doi}
\alias{check_doi}
\title{Check if a DOI exists}
\usage{
check_doi(doi)
}
\arguments{
\item{doi}{a \code{character} of length 1. A single DOI to check.}
}
\value{
A \code{logical}. If the DOI has the good structure and exists on
\url{https://www.doi.org/}, the returned value will be \code{TRUE}.
}
\description{
Checks if a DOI (Digital Object Identifier) has a good structure, i.e.
\verb{10.----/suffix}. Also checks if the redirection using the \strong{DOI System}
(\url{https://www.doi.org/}) works. An internet connection is required.
}
\examples{
check_doi("bad_doi")
check_doi("10.1016/j.jviromet.2032.01.007")
check_doi("10.1016/j.jviromet.2008.01.007")
}
