% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-title.R
\name{clean_title}
\alias{clean_title}
\title{Clean a title extracted from a citation}
\usage{
clean_title(x)
}
\arguments{
\item{x}{a \code{character} of length 1. A title extracted from a citation to
clean.}
}
\value{
A \code{character} of length 1 with the extracted title. If the title
cannot be detected, an empty string is returned.
}
\description{
Cleans a string by removing numbers, caesuras, punctuation, multiple
whitespace, leading, and trailing whitespace.
}
\examples{
ref <- paste0("Quainoo, A.K., Wetten, A.C., Allainguillaume, J., 2008. The ",
              "effectiveness of somatic embryogenesis in eliminating the ", 
              "cocoa swollen shoot virus from infected co- coa trees. ", 
              "J. Virol. Methods 149, 91e96.")
ref

ref_title <- title_from_citation(ref)
ref_title

clean_title(ref_title)
}
