% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi-to-bib.R
\name{doi_to_bib}
\alias{doi_to_bib}
\title{Convert DOI to BiBTeX}
\usage{
doi_to_bib(dois, filename)
}
\arguments{
\item{dois}{a \code{character} vector. A vector of DOI.}

\item{filename}{a \code{character} of length 1. The name of the \code{.bib} file
(with file extension) to create.}
}
\value{
BiBTeX entries with DOI.
}
\description{
From a vector of DOI, creates a BiBTeX expression with only DOI as metadata.
No additional metadata (authors, title, etc.) are added. The type of document
is set to \code{article} and a fictitious BiBTeX identifier is created.

The resulting \code{.bib} file can be imported in a reference management software
(Mendeley, Zotero, etc.) to retrieve complete metadata/full text.
}
\examples{
## Temporary file ----
file_name <- tempfile()

dois <- c("10.1080/03650340.2012.706871", 
          "10.1111/j.1744-7348.1970.tb05510.x")
          
bib_entries <- doi_to_bib(dois, file_name)
cat(bib_entries)
}
