% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve article metadata from a DOI}
\usage{
get_metadata(doi)
}
\arguments{
\item{doi}{a character of length 1. The Digital Object Identifier (DOI) of
the publication fo which metadata must be retrieved.}
}
\value{
A \code{data.frame} with publication metadata. If no match in WOS, the
\code{data.frame} will be empty.
}
\description{
Retrieves publication metadata from a DOI by querying Web of Science Lite
API. See \code{\link[=search_in_wos]{search_in_wos()}} for further information.
}
\examples{
\dontrun{
get_metadata("10.1111/ele.13778")
}
}
