% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pdf.R
\name{get_pdf}
\alias{get_pdf}
\title{Download article PDF from Sc.-H.b}
\usage{
get_pdf(doi, download = TRUE, path = ".", filename = "", overwrite = FALSE)
}
\arguments{
\item{doi}{a \code{character} of length 1. The DOI of the article.}

\item{download}{a \code{logical}. If \code{TRUE} (default) the PDF is downloaded.
Otherwise the link to the PDF is returned.}

\item{path}{a \code{character} of length 1. The directory to save PDFs.
Must exist.}

\item{filename}{a \code{character} of length 1. The name of PDF
(without extension).}

\item{overwrite}{a \code{logical}. If \code{TRUE} and the PDF already exist, the PDF
will be downloaded again. Default if \code{FALSE}.}
}
\value{
The PDF link (\code{character} of length 1) or \code{NULL} (no PDF available).
}
\description{
Downloads article PDF from the famous Sc.-H.b website. The URL of the website
must be stored locally as the environment variable \code{SCIHUB_URL} in the
\verb{~/.Renviron}. Use the function \code{\link[usethis:edit]{usethis::edit_r_environ()}} to edit this
file.
}
\examples{
\dontrun{
doi <- "10.1111/ele.13778"
get_pdf(doi, path = ".", filename = "2022-mouillot-ecolet")
}
}
