% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_doi.R
\name{search_doi}
\alias{search_doi}
\title{Find the DOI of a citation using the Web of Science Lite and CrossRef APIs}
\usage{
search_doi(x, limit = 1)
}
\arguments{
\item{x}{a \code{character} of length 1. A citation written as
"Authors Year Title. Journal...". See \code{\link{title_from_citation}} for further
detail.}

\item{limit}{an \code{integer} of length 1. The maximum number of records to test
to identify the match. Must be strictly positive and < 100,0000. Only used
with WOS Lite API.}
}
\value{
A \code{data.frame} with one row and the following 5 columns:
\itemize{
\item \code{reference}: the original citation;
\item \code{search_term}: the cleaned title extracted from the citation used as the
query terms in APIs;
\item \code{best_title}: the title of the best reference returned by WOS/CrossRef;
\item \code{best_doi}: the DOI corresponding to the best reference returned by
WOS/CrossRef;
\item \code{string_dist}: the string distance (Optimal string aligment, \code{osa})
between the original title and the best match.
See \code{\link[stringdist:stringdist]{stringdist::stringdist}} for further information.
}
}
\description{
This function is a wrapper around \code{\link{search_in_wos}} and \code{\link{search_in_cr}}.

From a citation, extracts the title and searches the corresponding
DOI on the Web of Science database using the WOS Lite API
(\url{https://developer.clarivate.com/apis/woslite}) and the R package
\code{rwoslite} (\url{https://github.com/frbcesab/rwoslite}). If the WOS Lite API
does not return any match, then the search is performed using the CrossRef
API
(\url{https://www.crossref.org/documentation/retrieve-metadata/rest-api/})
and the R package \code{rcrossref}
(\url{https://cran.r-project.org/package=rcrossref}).

\strong{Important}: for the WOS Lite API, a token must be first obtained and
stored locally. See \url{https://github.com/frbcesab/rwoslite} for further
information. For the CrossRef API, you need to share your email address with
the API. See
\url{https://docs.ropensci.org/rcrossref/#register-for-the-polite-pool}
for further information.

The extraction of the title is performed by the function
\code{\link{title_from_citation}} and may fail if the citation is malformed.
}
\examples{
\dontrun{
ref <- paste0("Quainoo, A.K., Wetten, A.C., Allainguillaume, J., 2008. The ",
              "effectiveness of somatic embryogenesis in eliminating the ", 
              "cocoa swollen shoot virus from infected co- coa trees. ", 
              "J. Virol. Methods 149, 91e96.")
              
search_doi(ref)
}
}
