% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-in-cr.R
\name{search_in_cr}
\alias{search_in_cr}
\title{Find the DOI of a citation using the CrossRef API}
\usage{
search_in_cr(x)
}
\arguments{
\item{x}{a \code{character} of length 1. A citation written as
"Authors Year Title. Journal...". See \code{\link{title_from_citation}} for further
detail.}
}
\value{
A \code{data.frame} with one row and the following 4 columns:
\itemize{
\item \code{search_term}: the cleaned title extracted from the citation used as the
query terms in CrossRef API;
\item \code{cr_best_title}: the title of the best reference returned by CrossRef;
\item \code{cr_best_doi}: the DOI corresponding to the best reference returned by
CrossRef;
\item \code{cr_string_dist}: the string distance (Optimal string aligment, \code{osa})
between the original title and the one returned by CrossRef.
See \code{\link[stringdist:stringdist]{stringdist::stringdist}} for further information.
}
}
\description{
From a citation, extracts the title and searches the corresponding
DOI on the CrossRef database using the CrossRef API
(\url{https://www.crossref.org/documentation/retrieve-metadata/rest-api/})
and the R package \code{rcrossref}
(\url{https://cran.r-project.org/package=rcrossref}).

\strong{Important}: you need to share your email address with Crossref. See
\url{https://docs.ropensci.org/rcrossref/#register-for-the-polite-pool}
for further information.

The extraction of the title is performed by the function
\code{\link{title_from_citation}} and may fail if the citation is malformed.

The candidate title (returned by the CrossRef API) is compared to the
original title, and the string distance is performed using the function
\code{\link[stringdist:stringdist]{stringdist::stringdist}} to have an idea of the match.
}
\examples{
\dontrun{
ref <- paste0("Quainoo, A.K., Wetten, A.C., Allainguillaume, J., 2008. The ",
              "effectiveness of somatic embryogenesis in eliminating the ", 
              "cocoa swollen shoot virus from infected co- coa trees. ", 
              "J. Virol. Methods 149, 91e96.")
              
search_in_cr(ref)
}
}
