% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-from-citation.R
\name{title_from_citation}
\alias{title_from_citation}
\title{Extract the title from a citation}
\usage{
title_from_citation(x)
}
\arguments{
\item{x}{a \code{character} of length 1. A citation written as
"Authors Year Title. Journal..."}
}
\value{
A \code{character} of length 1 with the extracted title. If the title
cannot be detected, an empty string is returned.
}
\description{
The title must follow the year. This function detects different forms of
years (e.g. \code{2021 }, \code{2021.}, \code{(2021)}) and handles deduplicated years
(e.g. \verb{2021a}, \verb{2021b}, etc.).

The title must end with a dot. If the title contains internal dots, only
the part of the title left to the first dot will be returned.
}
\examples{
ref <- paste0("Quainoo, A.K., Wetten, A.C., Allainguillaume, J., 2008. The ",
              "effectiveness of somatic embryogenesis in eliminating the ", 
              "cocoa swollen shoot virus from infected co- coa trees. ", 
              "J. Virol. Methods 149, 91e96.")
ref

title_from_citation(ref)
}
