% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-title.R
\name{clean_title}
\alias{clean_title}
\title{Clean a title extracted from a reference citation}
\usage{
clean_title(x)
}
\arguments{
\item{x}{a \code{character} of length 1. A title extracted from a reference
citation to clean.}
}
\value{
A \code{character} of length 1 with the extracted title. If the title
cannot be detected, an empty string is returned.
}
\description{
Cleans a string by removing numbers, caesuras, punctuation, multiple
whitespace, leading, and trailing whitespace.
}
\examples{
ref <- paste("R Core Team (2022) R: A language and environment for ", 
             "statistical computing. R Foundation for Statistical ",
             "Computing, Vienna, Austria. URL https://www.R-project.org/.")
ref

ref_title <- title_from_citation(ref)
clean_title(ref_title)
}
